\name{countries_list}
\alias{countries_list}
\docType{data}
\title{
Countries list
}
\description{
This file contains a list of all the countries in the world and its respective country code. You need the country codes (e.g. chl) to obtain data for a certain country (e.g Chile)
}
\usage{countries_list}
\format{
  A data frame with 305 observations on the following 2 variables.
  \describe{
    \item{\code{country}}{the full names of the countries}
    \item{\code{country_code}}{the ids of the countries}
  }
}
\examples{
countries_list
}
\keyword{datasets}
