\name{country_codes}
\alias{country_codes}
\title{Country codes}
\docType{data}
\description{
A reference to know the country codes to be able to download and import data (e.g. \code{chl} stands for "Chile" in the OEC's API). This follows the standard ISO-3 country codes.
}
\usage{country_codes}
\format{
  A data frame with 263 observations on the following 2 variables.
  \describe{
    \item{\code{country}}{Official countries' names spelled according to the United Nations.}
    \item{\code{country_code}}{Three characters codes to refer to the countries in the OEC's API.}
  }
}
\examples{
country_codes
}
\keyword{datasets}
