\name{d3plus-1.9.8.zip}
\alias{d3plus-1.9.8.zip}

\title{D3Plus visualization library}

\description{D3Plus is an extension to the D3 library that allows fast and easy creation of data visualizations.

This zip will be decompressed in your working directory when you use \code{\link{treemap}}, \code{\link{treemap_interval}}, \code{\link{network}} or \code{\link{network_interval}} for the first time in your working directory. It contains both the javascript and the icons required to display the visualizations in the browser.

D3Plus was created by Alexander Simoes and Dave Landry and D3 was created by Mike Bostock.}
