% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d3plus.R
\name{d3plus}
\alias{d3plus}
\title{Install D3 and D3plus in your working directory.}
\usage{
d3plus()
}
\value{
Copies a folder named \code{d3plus} to the working directory and it contains the \code{js} files and icons to make the visualizations
}
\description{
Install D3 and D3plus in your working directory.
}
\examples{
d3plus()
}
\keyword{functions}

