% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demos.R
\name{demos}
\alias{demos}
\title{Copies demo file with examples.}
\usage{
demos()
}
\value{
A file named \code{demo_examples.R} will be copied to the working directory.
}
\description{
Copies demo file with examples.
}
\examples{
demos()
}
\keyword{functions}

