\name{edges_sitc_4.json}
\alias{edges_sitc_4.json}

\title{A part of a pre-drawn network to create network visualizations using SITC (rev.2) product classification.}

\description{The network visualization is a bit different from the rest of D3Plus in that it requires extra files besides the data and attribute lookups. It is important to note that D3Plus network visualizations code does not attempt to create a dynamic layout based on a nodes and edges list.

This package provides a curated network provided in two JSON files of (X,Y) coordinates divided between elements positioning (nodes) and elements connections (edges). The Product Space described in the OEC website is a network of around 800 nodes and roughly 2000 edges, and this file is a part of a precomputed layout of the network that allows the user to display this large network without pushing all of the resources on the client's machine.}
