% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_countrycode.R
\name{get_countrycode}
\alias{get_countrycode}
\title{aaa}
\usage{
get_countrycode(countryname = "world")
}
\arguments{
\item{countryname}{aaa}
}
\value{
aaa
}
\description{
aaa
}
\examples{
get_countrycode("world")
}
\keyword{functions}
