% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Downloads and processes the data from the API}
\usage{
get_data(origin = "all", destination = "all", years = 2000,
  classification = "sitc")
}
\arguments{
\item{origin}{ISO code for country of origin (e.g. \code{chl} for Chile). Default set to \code{all}.
Run \code{country_codes} in case of doubt.}

\item{destination}{ISO code for country of destination (e.g. \code{chn} for China). Default set to \code{all}.
Run \code{country_codes} in case of doubt.}

\item{years}{Numeric value greater or equal to 1962 and lower of equal to 2016. Default set to 2000.}

\item{classification}{Any of the available trade classifications in the OEC (\code{sitc}, \code{hs92},
\code{hs96}, \code{hs02} or \code{hs07}). Default set to \code{sitc}.}
}
\value{
A tibble that describes bilateral trade metrics (imports, exports, trade balance and relevant metrics
such as exports growth w/r to last year) between an \code{origin} and \code{destination} country.
}
\description{
This function accesses \code{atlas.media.mit.edu}and perfoms different API calls to return tidy data.
and data transforming.
}
\examples{
\dontrun{
# The next examples can take more than 5 seconds to compute, and specially for large economies so
# these are just shown without evaluation according to CRAN rules

# Run `country_codes` to display the full table of countries

# What does Chile export to China?
# year 2015 - SITC (4 characters)
get_data("chl", "chn", 2015)
# or with explicit parameter
get_data("chl", "chn", 2015, "sitc")

# What does Chile export to China?
# years 2010, 2011 and 2015 - HS07 (4 and 6 characters)
get_data("chl", "chn", c(2010, 2011, 2015), "hs07")
}
}
\keyword{functions}
