% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_productcode.R
\name{get_productcode}
\alias{get_productcode}
\title{aaa}
\usage{
get_productcode(productname = "animals")
}
\arguments{
\item{productname}{aaa}
}
\value{
aaa
}
\description{
aaa
}
\examples{
get_productcode("animals")
}
\keyword{functions}
