% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata.R
\name{getdata}
\alias{getdata}
\title{Downloads and processes the data from the API}
\usage{
getdata(origin, destination, year, classification)
}
\arguments{
\item{origin}{Country code of origin (e.g. "chl" for Chile)}

\item{destination}{Country code of destination (e.g. "chn" for China)}

\item{year}{The OEC's API ranges from 1962 to 2014}

\item{classification}{Trade classification that can be "1" (HS92 4 characters since year 1995), "2" (SITC rev.2 4 characters since year 1962) or "3" (HS92 6 characters since year 1995)}
}
\description{
Downloads and processes the data from the API
}
\examples{
# Run countries_list() to display the full list of countries
# Chile is "chl" and China is "chn"

# Download trade data from OEC's API (HS92 4 characters product list)
# for Chile and China in the year 2014
# getdata("chl", "chn", 2014)
# is the same as
# getdata("chl", "chn", 2014, 1)

# Download trade data from OEC's API (SITC rev.2 4 characters product list)
# for Chile and China in the year 2014
# getdata("chl", "chn", 2014, 2)

# Download trade data from OEC's API (HS92 6 characters product list)
# for Chile and China in the year 2014
# getdata("chl", "chn", 2014, 3)
}
\keyword{functions}

