\name{hs92_6char}
\alias{hs92_6char}
\title{HS92 products (6 characters)}
\docType{data}
\description{
HS92 products (6 characters). This file is used to create the visualizations.
}
\usage{hs92_6char}
\format{
  A data frame with 1242 observations on the following 4 variables.
  \describe{
    \item{\code{hs92_product_name}}{Contains the H292 products' names (e.g. horses, bovine, pigs, etc)}
    \item{\code{hs92_group_name}}{Contains the H292 groups (e.g. animal products, vegetable products, etc)}
    \item{\code{hs92_product_id}}{Contains the associated codes of every product (e.g. horses is 010101)}
    \item{\code{hs92_group_id}}{Contains the associated codes of every group (e.g. animal products is 01)}
  }
}
\examples{
# see the group codes and product codes for HS92 (6 characters)
# hs92_6char
}
\keyword{datasets}
