\name{hs92_8char}
\alias{hs92_8char}
\title{HS92 products (8 characters)}
\docType{data}
\description{
HS92 products (8 characters). This file is used to create the visualizations.
}
\usage{hs92_8char}
\format{
  A data frame with 5040 observations on the following 4 variables.
  \describe{
    \item{\code{product}}{Contains the H292 products' names (e.g. horses, bovine, pigs, etc)}
    \item{\code{group}}{Contains the H292 groups (e.g. animal products, vegetable products, etc)}
    \item{\code{product_id}}{Contains the associated codes of every product (e.g. horses is 010101)}
    \item{\code{group_id}}{Contains the associated codes of every group (e.g. animal products is 01)}
  }
}
\examples{
hs92_8char
}
\keyword{datasets}
