\name{hs92_colors}
\alias{hs92_colors}
\title{HS92 colors}
\docType{data}
\description{
HS92 colors. This file is used to create the visualizations.
}
\usage{hs_colors}
\format{
  A data frame with 21 observations on the following 2 variables.
  \describe{
    \item{\code{hs92_group_name}}{Contains the H292 groups (e.g. animal products, vegetable products, etc)}
    \item{\code{hs92_color}}{Contains the associated colors of every group (e.g. mineral products is #330000)}
  }
}
\examples{
# see the group codes and group colors for HS92 (6 and 8 characters)
# hs92_colors
}
\keyword{datasets}
