\name{hs_colors}
\alias{hs_colors}
\docType{data}
\title{
HS colors
}
\description{
This file contains the HS92 colors. This file is used to create spreadsheets and visualizations with trade data.
}
\usage{hs_colors}
\format{
  A data frame with 21 observations on the following 2 variables.
  \describe{
    \item{\code{group}}{Contains the H292 groups (e.g. animal products, vegetable products, etc)}
    \item{\code{color}}{Contains the associated colors of every group (e.g. mineral products is #330000)}
  }
}
\examples{
hs_colors
}
\keyword{datasets}
