\name{network_template.html}
\alias{network_template.html}

\title{A template to display a network of the exports of a country for a certain year using HS92 or SITC (rev.2) product classification.}

\description{Contains a formatted template with fields that \code{\link{network}} or \code{\link{network_interval}} functions will find and replace accordingly to the data you want to display and the final visualization will be saved in an HTML file to your working directory.}
