% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oec_package.R
\name{oec-exports}
\alias{oec-exports}
\title{oec exported operators and S3 methods
The following functions are imported and then re-exported
from the oec package to avoid listing Depends of oec}
\description{
oec exported operators and S3 methods
The following functions are imported and then re-exported
from the oec package to avoid listing Depends of oec
}
