\name{sitc_colors}
\alias{sitc_colors}
\docType{data}
\title{
SITC (rev. 2) colors
}
\description{
This file contains the SITC (rev. 2) colors. This file is used to create spreadsheets and visualizations with trade data.
}
\usage{hs_colors}
\format{
  A data frame with 36 observations on the following 2 variables.
  \describe{
    \item{\code{group}}{Contains the SITC (rev. 2) groups (e.g. machinery, electronics, etc)}
    \item{\code{color}}{Contains the associated colors of every group (e.g. machinery is #17bcef)}
  }
}
\examples{
sitc_colors
}
\keyword{datasets}
