\name{sitc_rev2_4char}
\alias{sitc_rev2_4char}
\title{SITC (rev.2) products (4 characters)}
\docType{data}
\description{
SITC (rev. 2) products (4 characters). This file is used to create the visualizations.
}
\usage{sitc_rev2_4char}
\format{
  A data frame with 988 observations on the following 5 variables.
  \describe{
    \item{\code{sitc_rev2_product_name}}{Contains the SITC rev.2 products' names (e.g. initiating devices, polymerization ion exchangers, etc)}
    \item{\code{sitc_rev2_group_name}}{Contains the SITC rev.2 groups (e.g. machinery, electronics products, etc)}
    \item{\code{sitc_rev2_product_id}}{Contains the associated codes of every product (e.g. initiating devices is 5722)}
    \item{\code{sitc_rev2_group_id}}{Contains the associated codes of every group (e.g. machinery is 10)}
  }
}
\examples{
# see the group codes and product codes for SITC rev.2 (4 characters)
# sitc_rev2_4char
}
\keyword{datasets}
