\name{treemap_template.html}
\alias{treemap_template.html}

\title{A template to display a treemap of the imports, exports or trade balance of a country for a certain year using HS92 or SITC (rev.2) product classification.}

\description{Contains a formatted template with fields that \code{\link{treemap}} or \code{\link{treemap_interval}} functions will find and replace accordingly to the data you want to display and the final visualization will be saved in an HTML file to your working directory.}
