\name{countries_list}
\alias{countries_list}
\title{Countries list}
\docType{data}
\description{
A reference to know the country codes to be able to download and import data (e.g. \code{chl} stands for "Chile" in the OEC's API)
}
\usage{countries_list}
\format{
  A data frame with 263 observations on the following 2 variables.
  \describe{
    \item{\code{country}}{Official countries' names spelled according to the United Nations.}
    \item{\code{country_code}}{Three characters codes to refer to the countries in the OEC's API.}
  }
}
\examples{
countries_list
}
\keyword{datasets}
