\name{countries_list}
\alias{countries_list}
\title{A list of all the countries}
\docType{data}
\description{
A list of all the countries in the world and its country code. You need the country code (e.g. chl) to obtain data of a country (e.g Chile)
}
\usage{countries_list}
\format{
  A data frame with 305 observations on the following 2 variables.
  \describe{
    \item{\code{country}}{the full names of the countries}
    \item{\code{country_code}}{the ids of the countries}
  }
}
\examples{
countries_list
}
\keyword{datasets}
