% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demos.R
\name{demos}
\alias{demos}
\title{Copies the demo file}
\usage{
demos()
}
\value{
Copies a file named \code{demo_examples.R} to the working directory.
}
\description{
Copies the demo file
}
\examples{
demos()
}
\keyword{functions}

