% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata.R
\name{getdata}
\alias{getdata}
\title{Downloads and processes the data from the API}
\usage{
getdata(origin, dest, year, classification, write)
}
\arguments{
\item{origin}{Country code of origin (e.g. "chl" for Chile)}

\item{dest}{Country code of destination (e.g. "chn" for China)}

\item{year}{The OEC's API ranges from 1962 to 2016}

\item{classification}{Trade classification that can be "1" (HS92 4 characters since year 1995), "2" (SITC rev.2 4 characters since year 1962) or "3" (HS92 6 characters since year 1995)}

\item{write}{Write to user's filespace (by default set to "F")}
}
\description{
Downloads and processes the data from the API
}
\examples{
# Run countries_list to display the full list of countries

# What does Andorra export? (2015, HS92 4 characters)
getdata("and", "all", 2015)

# What does Andorra export? (2015, SITC rev2 4 characters)
getdata("and", "all", 2015, 2)

}
\keyword{functions}
