\name{hs92_6char}
\alias{hs92_6char}
\docType{data}
\title{
HS92 products (6 characters)
}
\description{
This file contains the HS92 products. This file is used to create spreadsheets and visualizations with trade data.
}
\usage{hs92_6char}
\format{
  A data frame with 1242 observations on the following 4 variables.
  \describe{
    \item{\code{product}}{Contains the H292 products' names (e.g. horses, bovine, pigs, etc)}
    \item{\code{group}}{Contains the H292 groups (e.g. animal products, vegetable products, etc)}
    \item{\code{product_id}}{Contains the associated codes of every product (e.g. horses is 010101)}
    \item{\code{group_id}}{Contains the associated codes of every group (e.g. animal products is 01)}
  }
}
\examples{
hs92_6char
}
\keyword{datasets}
