% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_interval.R
\name{network_interval}
\alias{network_interval}
\title{Creates a network of exports for a given year}
\usage{
network_interval(origin, destination, initial_year, final_year, interval,
  classification)
}
\arguments{
\item{origin}{is the country code of origin (e.g. "chl" for Chile)}

\item{destination}{is the country code of origin (e.g. "chn" for China)}

\item{classification}{refers to the trade classification that can be "6" (HS92 6 characters) or "8" (HS92 8 characters) for the year 1995 and going or "4" (SITC rev.2 4 characters) for the year 1962 and ongoing}

\item{year}{is the year and the OEC's API ranges from 1962 to 2014}
}
\value{
Creates an \code{HTML} file with a network visualization for a given year.
}
\description{
Creates a network of exports for a given year
}
\examples{
# Visualize trade data from OEC's API (HS92 6 characters product list)
# for exports from Chile to China in the year 2014
# network("chl", "chn", 2014, 6)
# is the same as
# network("chl", "chn", 2014)
}
\keyword{functions}

