% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_interval.R
\name{network_interval}
\alias{network_interval}
\title{Creates a network of exports for a given year}
\usage{
network_interval(origin, destination, initial_year, final_year, interval,
  classification)
}
\arguments{
\item{origin}{is the country code of origin (e.g. "chl" for Chile)}

\item{destination}{is the country code of origin (e.g. "chn" for China)}

\item{initial_year}{The OEC's API ranges from 1962 to 2014. This needs to be lower than `final_year`}

\item{final_year}{The OEC's API ranges from 1962 to 2014. This needs to be greater than `initial_year`}

\item{interval}{is an optional parameter to define the distance between years (by default set to 1)}

\item{classification}{Trade classification that can be "1" (HS92 4 characters since year 1995) or "2" (SITC rev.2 4 characters since year 1962)}
}
\value{
Creates an \code{HTML} file with a network visualization for a given year.
}
\description{
Creates a network of exports for a given year
}
\examples{
# Run countries_list() to display the full list of countries
# Chile is "chl" and China is "chn"
# Visualize trade data from OEC's API (HS92 4 characters product list)
# for exports from Chile to China in the year 2014
# network("chl", "chn", 2014, 1)
# is the same as
# network("chl", "chn", 2014)
}
\keyword{functions}

