% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemap.R
\name{treemap}
\alias{treemap}
\title{Creates a treemap for a given year}
\usage{
treemap(origin, destination, variable, year, classification, depth)
}
\arguments{
\item{origin}{is the country code of origin (e.g. "chl" for Chile)}

\item{destination}{is the country code of origin (e.g. "chn" for China)}

\item{variable}{is the variable to visualize and it can be "imports", "exports" or "exchange" (trade exchange)}

\item{year}{is the year and the OEC's API ranges from 1962 to 2014}

\item{classification}{refers to the trade classification that can be "6" (HS92 6 characters) or "8" (HS92 8 characters) for the year 1995 and going or "4" (SITC rev.2 4 characters) for the year 1962 and ongoing}

\item{depth}{is an optional parameter that can take values "0" (group's detail) or "1" (product's detail)}
}
\value{
Creates an \code{HTML} file with a treemap visualization for a given year.
}
\description{
Creates a treemap for a given year
}
\examples{
# Run countries_list() to display the full list of countries
# Chile is "chl" and China is "chn"
# Visualize trade data from OEC's API (HS92 6 characters product list)
# for Chile and China in the year 2014
# treemap("chl", "chn", "exports", 2014, 6)
# is the same as
# treemap("chl", "chn", "exports", 2014)
}
\keyword{functions}

